update Andre
Hi,

please have a look at my suggestions and see if the code works well till the end. I like the "mono" family for the plot, if you can keep that that's nice. Also I added the bounding box to determine the clipping/cropping processes to make it more generalisable to other datasets.
Thanks a lot. Cheers
Andre